Removing users who have left
Removing superadmins who have left.  In the early days we kept some of
these on as superadmins as we had few team members and were weren't sure
if some were returning.  Removing now as we have plenty of people in the
team and they are not returning to the platform.